Move data route configuration into the jpa autoconfig - SubscriptionOperationJpaAutoConfiguration
