+++ services/src/main/java/com/broadleafcommerce/subscriptionoperation/service/autoconfigure/SubscriptionOperationServiceAutoConfiguration.java
@@ -21,17 +21,12 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.lang.Nullable;
 import org.springframework.web.reactive.function.client.WebClient;
 
 import com.broadleafcommerce.common.extension.TypeFactory;
-import com.broadleafcommerce.common.extension.data.DataRouteSupporting;
-import com.broadleafcommerce.common.extension.data.PackageDataRouteSupplier;
-import com.broadleafcommerce.data.tracking.core.context.ContextInfoCustomizer;
 import com.broadleafcommerce.subscriptionoperation.domain.Subscription;
 import com.broadleafcommerce.subscriptionoperation.domain.SubscriptionItem;
 import com.broadleafcommerce.subscriptionoperation.domain.SubscriptionWithItems;
-import com.broadleafcommerce.subscriptionoperation.provider.jpa.environment.RouteConstants;
 import com.broadleafcommerce.subscriptionoperation.service.DefaultSubscriptionOperationService;
 import com.broadleafcommerce.subscriptionoperation.service.SubscriptionOperationService;
 import com.broadleafcommerce.subscriptionoperation.service.provider.SubscriptionProvider;
@@ -39,7 +34,6 @@ import com.broadleafcommerce.subscriptionoperation.service.provider.external.Ext
 import com.broadleafcommerce.subscriptionoperation.service.provider.external.ExternalSubscriptionProvider;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-import lombok.RequiredArgsConstructor;
 
 @Configuration
 public class SubscriptionOperationServiceAutoConfiguration {
@@ -68,42 +62,4 @@ public class SubscriptionOperationServiceAutoConfiguration {
                     properties);
         }
     }
-
-    /**
-     * Defines a {@link DataRouteSupporting} for Subscription Ops. By default, this is detached from
-     * any persistence and is used for supporting {@link ContextInfoCustomizer} in a flexpackage
-     * configuration to ensure it is only invoked for this specific service.
-     */
-    @Bean(name = "subscriptionOperationSource")
-    @ConditionalOnMissingBean(name = "subscriptionOperationSource")
-    public DataRouteSupporting subscriptionOperationSource() {
-        return new OrchestrationDataRouteSupporting(
-                RouteConstants.Persistence.SUBSCRIPTION_OPS_ROUTE_KEY);
-    }
-
-    @Bean(name = "subscriptionOperationRouteSupplier")
-    @ConditionalOnMissingBean(name = "subscriptionOperationRouteSupplier")
-    public PackageDataRouteSupplier<DataRouteSupporting> subscriptionOperationRouteSupplier(
-            @Nullable @Qualifier("subscriptionOperationSource") DataRouteSupporting route) {
-        return () -> new PackageDataRouteSupplier.PackageMapping<>(
-                RouteConstants.Persistence.SUBSCRIPTION_OPS_ROUTE_PACKAGE,
-                route);
-    }
-
-    @RequiredArgsConstructor
-    private static class OrchestrationDataRouteSupporting implements DataRouteSupporting {
-
-        private final String routeKey;
-
-        @Override
-        public String getLookupKey() {
-            return routeKey;
-        }
-
-        @Override
-        public int getOrder() {
-            return 0;
-        }
-    }
-
 }
